--- src/algorithm/blueprint/index.ts
@@ -1,2 +1,0 @@
-export * from './parser';
-export * from './blueprint';
